Create main.py
To change the window attributes, window_width and window_height are what you're looking for. You change attributes of the rectangle with rect_width and rect_height. Finally, set the speed variable to an integer for whatever speed you would like. The lower the number, the higher the speed, since it slows down the while loop. Setting fullscreen to True makes all the variables automatically adjust to a 1920x1080 monitor for a fullscreen display. Thank you and have fun!